requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
SQLAlchemy==1.4.26
requests==2.26.0
python-jose==3.3.0
fastapi-utils==0.2.1